chore: move ocr parameters to ccip shared package
This moves the ocr parameters to the ccip shared package as the only
users of this struct.
